refactor(search): extract index-restore helpers from start.rs to fix line-cap budget
Move try_locate_moved_root and restore_one_index from start.rs into a new
sibling module start_restore.rs. start.rs was 2209 lines, exceeding its
frozen allowlist budget of 2197 after the OOM error-propagation edits landed
in the OOM (#954) changes. The extracted functions are the natural split point:
they are the leaf helpers in the restore call chain, have no deps on start.rs-
private state, and were already pub(crate). start.rs drops to 1892 lines
(budget 2197); start_restore.rs is 342 lines (under the 500-line cap).
line-cap CI gate: bash scripts/check_line_cap.sh → 0 violations.
